Huge rewrite
The last version of mysimplehomepage assumed that there would be a Jade
template for every page, which is just stupid. You also had to list all
the project files in the configuration file, which just becomes busy
work once the file count rises. To fix these mistakes, a rewrite was in
order...

New features:
- Pages can now use the same template.
- Every file in the static files, pages and templates directories are
  processed.
- The relative URLs of the static files can now be accessed inside the
  templates.
- Pages can now have their own metadata.