Added recursive directory traversal.
The builder now scans all the sub directories for files to
compile. Discarded 'category' in favor of 'to': the 'static' object in
templates will have all the files categorized based on the static files'
file extension. For example: paths to stylus compiled CSS file paths can
be found from static.css.

All the static files that have no compiler will be copied to the output
directory as they are.

The directory structure in pages and static files directories will be
the same in the output directory, so therefore 'outfilepath' in page
objects is now obsolete.